feat(android) feat(web): Add initial sources for Android app
The initial commit for the Android app (although with MVP).

The app is going to be a thin client for the Bot commands.
Similar to what the Web UI client is supposed to be,
but on the Android devices.

The app has the following hierarchical DB structure:
Source platform -> Account -> Org -> Module
-> Command -> Argument
Also the app will keep some local settings and the values
of the arguments entered by the user.
The source code uses the architecture components as
proposed by Google (e.g. Room, DAO, ViewModel,
Repository, LiveData, Retrofit, Hilt etc.)

The app already supports logging in and authorizing
by Discords, selecting the org and executing commands
for it, interpreted as plain text.
For the most commands it's still useless because they
require Discord mentioned that are yet to be supported
on the app side. However the commands that do not
require mentions already work fine.
Also, the !makeimage command works properly and
send the result pictures as Base64 strings to the Android
client. For that purpose the BaseMessage class was
improved to be able to send attachments and suggestions
in the response.
Calling back from Discord to the app is implemented
using the Android App link feature.

Forks of this Android project can be customized by
making different versions of the fork.properties file
based on the fork_template.properties.

Also made some improvements on the server side
to supports the client modifications:

- Added comments to the Web API section of the source.
- Got rid of implicit usages of the message's channel as
argument, where possible, since for the UI commands
there is no Discord channel.
- "Servers" in the Web API source code are renamed
to "orgs" to avoid confusion.
- Fixed the procedure of gathering the scanner types for
arguments, so this info gets properly sent to the clients.
- Web UI features are renamed to UI features, since
they now support Android as well.
- In the !makeimage command the style argument is
now skipped in the sequential scan, and the text argument
has the FullStringArgScanner now. This is to avoid
necessity or surrounding the text with quotes,
which was quite not obvious, especially on the phones..
- Permit remind and deny remind now have proper
aliases for the channes arguments.
- The !poll command now has a channel argument,
which (theoretically) should allow to use it from
UI clients.
- Other minor fixes.

+++ src/commands_discord/image/make-image-command.js
@@ -7,10 +7,13 @@
  */
 
 const DiscordCommand = require('../discord-command');
+const BaseMessageAttachment = require('../../components/base-message-attachment');
 const ObjectArgScanner = require('../../arg_scanners/object-arg-scanner');
 const SimpleArgScanner = require('../../arg_scanners/simple-arg-scanner');
+const FullStringArgScanner = require('../../arg_scanners/full-string-arg-scanner');
 const CommandArgDef = require('../../command_meta/command-arg-def');
 const BotPublicError = require('../../utils/bot-public-error');
+const jimp = require('jimp');
 const { v4: uuidv4 } = require('uuid');
 const fs = require('fs');
 
@@ -33,23 +36,24 @@ const MakeImageCommandArgDefs = Object.freeze({
     scanner: SimpleArgScanner,
     validationOptions: { nonNull: true }
   }),
-  text: new CommandArgDef('text', {
-    aliasIds: [
-      'command_makeimage_arg_text_alias_text',
-      'command_makeimage_arg_text_alias_t'
-    ],
-    helpId: 'command_makeimage_arg_text_help',
-    scanner: SimpleArgScanner,
-    validationOptions: { nonNull: true }
-  }),
   style: new CommandArgDef('style', {
     aliasIds: [
       'command_makeimage_arg_style_alias_style',
       'command_makeimage_arg_style_alias_s'
     ],
     helpId: 'command_makeimage_arg_style_help',
+    skipInSequentialRead: true,
     scanner: ObjectArgScanner,
     validationOptions: { nonNull: false }
+  }),
+  text: new CommandArgDef('text', {
+    aliasIds: [
+      'command_makeimage_arg_text_alias_text',
+      'command_makeimage_arg_text_alias_t'
+    ],
+    helpId: 'command_makeimage_arg_text_help',
+    scanner: FullStringArgScanner,
+    validationOptions: { nonNull: true }
   })
 });
 
@@ -153,18 +157,24 @@ class MakeImageCommand extends DiscordCommand {
     try {
       const imageResult = await this.context.imageGenerator.generateImage(this.imgUrl, params, this.parsedJsonConfig,
         this.source, this.orgId);
-      const filePath = `images/${uuidv4()}.jpg`;
-      imageResult.write(filePath);
-      await message.originalMessage.channel.send(null, {
-        files: [filePath]
-      });
 
-      await fs.unlink(filePath, (err) => {
-        if (err) {
-          this.context.log.e(`Failed to delete file ${filePath}. Error: ${err}`);
-        }
-      });
+      if (message.originalMessage.channel != null) {
+        const filePath = `images/${uuidv4()}.jpg`;
+        imageResult.write(filePath);
+        await message.originalMessage.channel.send(null, {
+          files: [filePath]
+        });
 
+        await fs.unlink(filePath, (err) => {
+          if (err) {
+            this.context.log.e(`Failed to delete file ${filePath}. Error: ${err}`);
+          }
+        });
+      } else {
+        const base64Image = await imageResult.getBase64Async(jimp.MIME_JPEG);
+        message.replyResult.attachments.push(new BaseMessageAttachment(base64Image,
+          BaseMessageAttachment.MIME_TYPES.BASE64));
+      }
     } catch (e) {
       this.context.log.e(this.name + ': Cannot make image: ' + e.stack);
       return this.langManager.getString('command_makeimage_error_broken_template', this.templateName, e.message);
